Separated project into 3 layers
