+++ Backend/CarRentalApp/CarRentalApp.sln
@@ -5,6 +5,10 @@ VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CarRentalApp", "CarRentalApp\CarRentalApp.csproj", "{1886D836-30ED-4896-B26D-4C1769FE8BCD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarRentalDal", "CarRentalDal\CarRentalDal.csproj", "{F57B80FE-120B-493B-8BE0-1A2D8D416D6B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CarRentalBll", "CarRentalBll\CarRentalBll.csproj", "{27ACE3C1-E0BC-490C-804B-15ACBFEAF848}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +19,14 @@ Global
 		{1886D836-30ED-4896-B26D-4C1769FE8BCD}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1886D836-30ED-4896-B26D-4C1769FE8BCD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1886D836-30ED-4896-B26D-4C1769FE8BCD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F57B80FE-120B-493B-8BE0-1A2D8D416D6B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F57B80FE-120B-493B-8BE0-1A2D8D416D6B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F57B80FE-120B-493B-8BE0-1A2D8D416D6B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F57B80FE-120B-493B-8BE0-1A2D8D416D6B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27ACE3C1-E0BC-490C-804B-15ACBFEAF848}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27ACE3C1-E0BC-490C-804B-15ACBFEAF848}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27ACE3C1-E0BC-490C-804B-15ACBFEAF848}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27ACE3C1-E0BC-490C-804B-15ACBFEAF848}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
