+++ Backend/CarRentalApp/CarRentalApp/Controllers/AuthController.cs
@@ -1,6 +1,6 @@
-using CarRentalApp.Models.Bll;
 using CarRentalApp.Models.Web.Requests;
-using CarRentalApp.Services;
+using CarRentalBll.Models;
+using CarRentalBll.Services;
 using Mapster;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
