+++ Backend/CarRentalApp/CarRentalApp/Controllers/UsersController.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Immutable;
-using CarRentalApp.Models.Bll;
 using CarRentalApp.Models.Web.Requests;
-using CarRentalApp.Services;
+using CarRentalBll.Models;
+using CarRentalBll.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
