+++ Backend/CarRentalApp/CarRentalApp/appsettings.json
@@ -42,6 +42,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "CarRentalDB": "Server=WSB-084-74\\MSSQSERVER1;Database=CarRentalDB;Trusted_Connection=True;"
+    "CarRentalDB1": "Server=WSB-084-74\\MSSQSERVER1;Database=CarRentalDB;Trusted_Connection=True;",
+    "CarRentalDB": "Server=DESKTOP-I1L3GQA\\MSSQLSERVER01;Database=CarRentalDB;Trusted_Connection=True;"
   }
 }
