+++ Backend/CarRentalApp/CarRentalBll/Configuration/JWT/Refresh/RefreshJwtConfig.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.IdentityModel.Tokens;
 
-namespace CarRentalApp.Configuration.JWT.Refresh
+namespace CarRentalBll.Configuration.JWT.Refresh
 {
     public class RefreshJwtConfig
     {
