+++ Backend/CarRentalApp/CarRentalBll/Configuration/JWT/Refresh/RefreshTokenGenerationParameters.cs
@@ -1,4 +1,4 @@
-﻿namespace CarRentalApp.Configuration.JWT.Refresh
+﻿namespace CarRentalBll.Configuration.JWT.Refresh
 {
     public class RefreshTokenGenerationParameters : GenerationParameters
     {
